memory: comment out _ACCESS_POOL in XRenderNULL and XRenderOGL
This fixes release builds of XRenderNULL and XRenderOGL
----- Additional info -----
_ACCESS_POOL could be declared as:

or as:
  _CryMemoryManagerPoolHelper g_CryMemoryManagerHelper;\
	void* CryModuleMalloc( size_t size ) throw(){ return g_CryMemoryManagerHelper.Malloc(size); }; \
...
(check SourceCode\CryCommon\CryMemoryManager.h)

In 1st case we didn't miss anything.
In 2nd case we just need to make sure that we have _ACCESS_POOL in
some other cpp file of current project. And we already have it in
SourceCode\RenderDll\Common\Renderer.cpp